+++ experiments/integrated/global/global_soil_canopy.jl
@@ -181,7 +181,7 @@ soil_model_type = Soil.EnergyHydrology{FT}
 soilco2_type = Soil.Biogeochemistry.SoilCO2Model{FT}
 
 # soil microbes args
-Csom = (z, t) -> eltype(z)(5.0)
+Csom = ClimaLand.PrescribedSoilOrganicCarbon{FT}(TimeVaryingInput((t) -> 5))
 
 # Set the soil CO2 BC to being atmospheric CO2
 soilco2_top_bc = Soil.Biogeochemistry.AtmosCO2StateBC()
@@ -368,7 +368,8 @@ sv = (;
 )
 saving_cb = ClimaLand.NonInterpSavingCallback(sv, saveat)
 updateat = Array(t0:(3600 * 3):tf)
-updatefunc = ClimaLand.make_update_drivers(atmos, radiation)
+drivers = ClimaLand.get_drivers(land)
+updatefunc = ClimaLand.make_update_drivers(drivers)
 driver_cb = ClimaLand.DriverUpdateCallback(updateat, updatefunc)
 cb = SciMLBase.CallbackSet(driver_cb, saving_cb)
 @time sol = SciMLBase.solve(