+++ experiments/standalone/Vegetation/varying_lai.jl
@@ -192,7 +192,8 @@ sv = (;
 saving_cb = ClimaLand.NonInterpSavingCallback(sv, saveat);
 
 updateat = Array(t0:1800:tf)
-updatefunc = ClimaLand.make_update_drivers(atmos, radiation)
+drivers = ClimaLand.get_drivers(canopy)
+updatefunc = ClimaLand.make_update_drivers(drivers)
 driver_cb = ClimaLand.DriverUpdateCallback(updateat, updatefunc)
 cb = SciMLBase.CallbackSet(driver_cb, saving_cb);